Move StorageMetrics unit tests into cpp file
