+++ fdbserver/RestoreLoader.actor.cpp
@@ -27,7 +27,7 @@
 #include "fdbserver/RestoreLoader.actor.h"
 #include "fdbserver/RestoreRoleCommon.actor.h"
 #include "fdbserver/MutationTracking.h"
-#include "fdbserver/StorageMetrics.actor.h"
+#include "fdbserver/StorageMetrics.h"
 
 #include "flow/actorcompiler.h" // This must be the last #include.
